third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/webnn-service-on-cpu/external/wpt/webnn/conformance_tests/gatherElements.https.any_cpu-expected.txt

This is a testharness.js-based test.
[FAIL] gatherElements float32 2D input and uint32 indices options.axis=1
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'indices' data type uint32 must be one of [int32]."
[FAIL] gatherElements float32 3D input and int32 negative indices
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float32 actual -66.05901336669922 should be close enough to expected 89.0337142944336 by ULP distance: expected a number less than or equal to 0n but got 2234920826n
[FAIL] gatherElements float32 1D input and int32 out-of-bounds indices
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float32 actual -26.158037185668945 should be close enough to expected 51.79948425292969 by ULP distance: expected a number less than or equal to 0n but got 2216719957n
Harness: the test ran to completion.
